number of pages to long

+++ src/main/java/com/quar17esma/dao/FeedbackDAO.java
@@ -5,7 +5,7 @@ import com.quar17esma.entity.Feedback;
 import java.util.List;
 
 public interface FeedbackDAO extends GenericDAO<Feedback> {
-    List<Feedback> findByPage(int page, int feedbackOnPage);
+    List<Feedback> findByPage(long page, int feedbackOnPage);
 
     long countAll();
 }
