+++ src/main/java/com/quar17esma/service/IFeedbackService.java
@@ -5,7 +5,7 @@ import com.quar17esma.entity.Feedback;
 import java.util.List;
 
 public interface IFeedbackService extends GenericService<Feedback> {
-    List<Feedback> getByPage(int page, int feedbacksOnPage);
+    List<Feedback> getByPage(long page, int feedbacksOnPage);
 
     long getAllQuantity();
 }
